refactor listings.go
